requirements.txt
solana==0.30.2
aiohttp==3.8.5
python-dotenv==1.0.0
asyncio==3.4.3
solders==0.18.1
